requirements.txt
matplotlib==2.1.1
joblib==0.15.1
numpy==1.13.3
streamlit==0.72.0
pandas==0.22.0
Pillow==8.0.1
scikit_learn==0.23.2
